fix circular import by moving some stuff around
