+++ vantage6-server/vantage6/server/permission.py
@@ -1,15 +1,52 @@
 import logging
 
+from flask import g
+from vantage6.server import db
+
 from vantage6.backend.common.permission import RuleCollectionBase, PermissionManagerBase
 from vantage6.backend.common.resource.error_handling import UnauthorizedError
 from vantage6.server.model.base import Base
 from vantage6.server.model.role import Role
 from vantage6.server.model.rule import Rule, Operation, Scope
-from vantage6.server.utils import obtain_auth_collaborations, obtain_auth_organization
 from vantage6.common import logger_name
 
 module_name = logger_name(__name__)
 log = logging.getLogger(module_name)
+
+
+def obtain_auth_collaborations() -> list[db.Collaboration]:
+    """
+    Obtain the collaborations that the auth is part of.
+
+    Returns
+    -------
+    list[db.Collaboration]
+        List of collaborations
+    """
+    if g.user:
+        return g.user.organization.collaborations
+    elif g.node:
+        return g.node.organization.collaborations
+    else:
+        return [db.Collaboration.get(g.container["collaboration_id"])]
+
+
+def obtain_auth_organization() -> db.Organization:
+    """
+    Obtain the organization model from the auth that is logged in.
+
+    Returns
+    -------
+    db.Organization
+        Organization model
+    """
+    if g.user:
+        org_id = g.user.organization.id
+    elif g.node:
+        org_id = g.node.organization.id
+    else:
+        org_id = g.container["organization_id"]
+    return db.Organization.get(org_id)
 
 
 class RuleCollection(RuleCollectionBase):
