+++ vantage6-server/vantage6/server/utils.py
@@ -1,38 +1,31 @@
-from flask import g
-
-from vantage6.server import db
+import datetime as dt
 
 
-def obtain_auth_collaborations() -> list[db.Collaboration]:
+def parse_datetime(
+    date: str | dt.datetime = None, default: dt.datetime = None
+) -> dt.datetime:
     """
-    Obtain the collaborations that the auth is part of.
+    Utility function to parse a datetime string.
+
+    Parameters
+    ----------
+    date : str | datetime.datetime, optional
+        Datetime string
+    default : datetime.datetime, optional
+        Default datetime to return if `dt` is None
 
     Returns
     -------
-    list[db.Collaboration]
-        List of collaborations
+    datetime.datetime
+        Datetime object
     """
-    if g.user:
-        return g.user.organization.collaborations
-    elif g.node:
-        return g.node.organization.collaborations
-    else:
-        return [db.Collaboration.get(g.container["collaboration_id"])]
-
-
-def obtain_auth_organization() -> db.Organization:
-    """
-    Obtain the organization model from the auth that is logged in.
-
-    Returns
-    -------
-    db.Organization
-        Organization model
-    """
-    if g.user:
-        org_id = g.user.organization.id
-    elif g.node:
-        org_id = g.node.organization.id
-    else:
-        org_id = g.container["organization_id"]
-    return db.Organization.get(org_id)
+    if date:
+        if isinstance(date, str):
+            converter = "%Y-%m-%dT%H:%M:%S.%f"
+            if date.endswith("+00:00"):
+                converter += "%z"  # parse timezone
+            return dt.datetime.strptime(date, converter)
+        else:
+            # convert datetime to UTC
+            return date.astimezone(dt.timezone.utc)
+    return default
